♻️ Changed op process's place
